+++ tracer/src/Datadog.Trace/Agent/StatsBucket.cs
@@ -35,10 +35,10 @@ namespace Datadog.Trace.Agent
 
         public long TopLevelHits { get; set; }
 
-        // Tracked for OTLP histogram data points; sentinels double.MaxValue/0 mean "not yet observed".
-        public double MinDuration { get; set; } = double.MaxValue;
+        // Tracked for OTLP histogram data points; sentinels long.MaxValue/0 mean "not yet observed".
+        public long MinDuration { get; set; } = long.MaxValue;
 
-        public double MaxDuration { get; set; }
+        public long MaxDuration { get; set; }
 
         public List<byte[]> PeerTags { get; }
 
@@ -50,7 +50,7 @@ namespace Datadog.Trace.Agent
             Errors = 0;
             Duration = 0;
             TopLevelHits = 0;
-            MinDuration = double.MaxValue;
+            MinDuration = long.MaxValue;
             MaxDuration = 0;
             OkSummary.Clear();
             ErrorSummary.Clear();
